requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
openai==1.3.8
python-dotenv==1.0.0
asyncio==3.4.3
logging==0.4.9.6
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# PDF обработка
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
tabula-py==2.8.2

# Дополнительные зависимости для качественной обработки
chardet==5.2.0
xlrd==2.0.1
matplotlib==3.8.2
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0